Do not parse and invoke callback when no response data is written in uncompressed mode
Often, a `res.write` will not be called. Currently in this scenario,
we still attempt to parse the written body (which is undefined), and
eventually attempt to convert it back into a Buffer. Both of these of
attemps fail.

I've refactored the `handleUncompressed` method to ensure we only
parse the response body and invoke the modify callback if any body has
been written. Otherwise, we simply just call `res.end` with no body data.